# 🚀 Major Performance Breakthrough: 97% Faster Pipeline + Critical Bug Fixes